Nasty bugfix for large packets
On sending a large packet, if the buffer filled up and we had to poll and wait for it to become available, we wouldn't adjust the buffer position based on the number of bytes already successfully sent, so we'd end up sending the *start* of the message again and repeating it from the beginning. This totally mangled the serialization on the client end. Switched the send and receive functions to accept uint8_t pointer types instead of void so we could do pointer arithmetic without casting.